调整商品管理

+++ src/main/java/com/example/ex3_2_back/controller/ProductController.java
@@ -24,19 +24,26 @@ public class ProductController {
     @GetMapping("/showList")
     @Operation(summary = "展示商品列表", description = "展示商品列表")
     public Result showList(){
-        return Result.success(productRepository.findAll());
+        if (productRepository.count()==0){
+            return Result.error("没有数据");
+        }else {
+            return Result.success(productRepository.findAll());
+        }
+
     }
 
     @GetMapping("/search")
     @Operation(summary = "查询商品信息", description = "查询商品信息")
     public Result search(@RequestParam("name") String name){
-        return Result.success(productRepository.findByName(name));
-    }
-
-    @GetMapping("/like")
-    @Operation(summary = "按名字模糊查询", description = "按名字模糊查询")
-    public Result like(@RequestParam("name") String name){
-        return Result.success(productRepository.findAllByNameLike("%"+name+"%"));
+        if (productRepository.existsByNameLike("%"+name+"%")){
+            if (productRepository.existsByName(name)){
+                return Result.success(productRepository.findByName(name));
+            }else {
+                return Result.success(productRepository.findAllByNameLike("%"+name+"%"));
+            }
+        }else {
+            return Result.error("未找到该商品或相似商品");
+        }
     }
 
     @PostMapping("/new")
@@ -46,40 +53,40 @@ public class ProductController {
             productRepository.save(product);
             return Result.success();
         }catch (Exception e){
-            return Result.error(e.getMessage()).addErrors(e);
+            return Result.error("商品"+"'"+product.getName()+"'"+"已存在");
         }
     }
 
     @DeleteMapping("/delete")
     @Operation(summary = "删除商品", description = "删除商品")
     public Result delete(@RequestParam("name") String name){
-        try {
+        if(productRepository.existsByName(name)){
             productRepository.deleteByName(name);
             return Result.success();
-        }catch (Exception e){
-            return Result.error(e.getMessage()).addErrors(e);
+        }else {
+            return Result.error("商品"+"'"+name+"'"+"不存在，请重新输入");
         }
     }
 
     @PutMapping("/update")
     @Operation(summary = "更新商品信息", description = "更新商品信息")
     public Result update(@RequestParam("description") String description, @RequestParam("price") BigDecimal price, @RequestParam("name") String name){
-        try {
+        if(productRepository.existsByName(name)){
             productRepository.updateInfoByName(description,price,name);
             return Result.success();
-        }catch (Exception e){
-            return Result.error(e.getMessage()).addErrors(e);
+        }else {
+            return Result.error("商品名不存在，请重试");
         }
     }
 
     @PutMapping("/listing")
     @Operation(summary = "商品上架和下架", description = "商品上架和下架")
     public Result listing(@RequestParam("isList") boolean isList, @RequestParam("name") String name){
-        try {
+        if(productRepository.existsByName(name)){
             productRepository.updateIsListByName(isList,name);
             return Result.success();
-        }catch (Exception e){
-            return Result.error(e.getMessage()).addErrors(e);
+        }else {
+            return Result.error("商品名不存在，请重试");
         }
     }
 }
